+++ findisco.cpp
@@ -1,22 +1,22 @@
 /* Calculate ISCO radius */
 // TODO: improve this with a more modular approach
 // TODO: optimise expressions
-// const long double Pi  = 3.14159265358979323846264338327950288419716939937510L;
+// const double Pi  = 3.14159265358979323846264338327950288419716939937510L;
 
-// long double spin = 0.1;
-// long double defpar = 0.005, epsi3 = 0.0, a13 = 0.0, a22 = 0.0;
+// double spin = 0.1;
+// double defpar = 0.005, epsi3 = 0.0, a13 = 0.0, a22 = 0.0;
 
 // #include <stdio.h>
 // #include <math.h>
 
 // #include "metric.cpp"
 
-long double find_isco()
+double find_isco()
 {
-    long double d2Veff, d2Veff2, E_var, Lz_var, Omega_var, denom, den, den_r, den_rr, num, num_r, num_rr;
-    long double m[4][4],dmdr[4][4],dmdr2[4][4];
-    long double l, j, r, d2Veff_last = 0, d2Veff_last2 = 1000, rin;
-    long double jmin, jmax, lmin, lmax, rmin, factor, rstep;
+    double d2Veff, d2Veff2, E_var, Lz_var, Omega_var, denom, den, den_r, den_rr, num, num_r, num_rr;
+    double m[4][4],dmdr[4][4],dmdr2[4][4];
+    double l, j, r, d2Veff_last = 0, d2Veff_last2 = 1000, rin;
+    double jmin, jmax, lmin, lmax, rmin, factor, rstep;
 
     rstep = 0.1;
     rmin = 1.1;
@@ -106,7 +106,7 @@ long double find_isco()
 }
 
 // int main(){
-//     long double isco;
+//     double isco;
 //     isco = find_isco();
 //     printf("ISCO radius = %Lf\n", isco);
 //     return 0;
